+++ src/Controller/Admin/BannersController.php
@@ -90,6 +90,10 @@ class BannersController extends AppController {
 		$this->paginate['order'] = ['Banners.created' => 'DESC'];
 		$this->paginate['sortWhitelist'] = ['Banners.filename', 'Positions.name', 'description', 'click_count', 'Banners.created'];
 		
+		//Sets the paginate limit and the maximum paginate limit
+		//See http://book.cakephp.org/3.0/en/controllers/components/pagination.html#limit-the-maximum-number-of-rows-that-can-be-fetched
+		$this->paginate['limit'] = $this->paginate['maxLimit'] = config('backend.photos');
+		
 		$this->set('banners', $this->paginate($this->Banners->queryFromFilter($query, $this->request->query)));
         
         if($render) {