+++ src/Controller/Admin/PhotosController.php
@@ -90,6 +90,10 @@ class PhotosController extends AppController {
 		$this->paginate['order'] = ['Photos.created' => 'DESC'];
 		$this->paginate['sortWhitelist'] = ['filename', 'Albums.title', 'Photos.created'];
 		
+		//Sets the paginate limit and the maximum paginate limit
+		//See http://book.cakephp.org/3.0/en/controllers/components/pagination.html#limit-the-maximum-number-of-rows-that-can-be-fetched
+		$this->paginate['limit'] = $this->paginate['maxLimit'] = config('backend.photos');
+		
 		$this->set('photos', $this->paginate($this->Photos->queryFromFilter($query, $this->request->query)));
         
         if($render) {